Singleton reference for cursor handle

+++ arch/x86_64/kernel/drivers/vga.cpp
@@ -19,7 +19,7 @@ cursor& cursor::operator<<(char c) {
     return *this;
 }
 
-cursor& cursor::operator<<(int n) {
+cursor& cursor::operator<<(size_t n) {
     char buffer[10];
     itoa(n, buffer, 10);
     return (*this) << buffer;
